move pathbuf structures
